third_party/blink/web_tests/platform/generic/editing/selection/caret-in-div-containing-br-in-vertical-mode-expected.txt

Testcase for bug 103621: Caret is incorrectly painted for a contenteditable <div> containing a <br> in vertical writing mode. For the test to pass, carets in both the boxes should be exactly similar. To manually verify the issue, try clicking on the second box. A caret should be visible at the start of that container.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS withTextCaretRect.top is withoutTextCaretRect.top
PASS withTextCaretRect.width is withoutTextCaretRect.width
PASS withTextCaretRect.height is withoutTextCaretRect.height
PASS successfullyParsed is true

TEST COMPLETE
